+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 
 - package `github.com/Azure/azure-sdk-for-go/services/billing/mgmt/2020-05-01/billing` has been renamed to `github.com/Azure/azure-sdk-for-go/services/preview/billing/mgmt/2020-05-01-preview/billing`
 
+- package `github.com/Azure/azure-sdk-for-go/services/preview/cdn/mgmt/2020-04-15/cdn` has been renamed to `github.com/Azure/azure-sdk-for-go/services/cdn/mgmt/2020-04-15/cdn`
+
 Generated from https://github.com/Azure/azure-rest-api-specs/tree/93106ff722d481c30a9a4ae089564635c10e9401
 
 ## `v46.4.0`